Adapt doc ID int types to placate -Wconversion.

+++ core/Lucy/Index/Posting/RawPosting.c
@@ -128,7 +128,7 @@ RawPostWriter_Write_Posting_IMP(RawPostingWriter *self, RawPosting *posting) {
     RawPostingIVARS *const posting_ivars = RawPost_IVARS(posting);
     OutStream *const outstream   = ivars->outstream;
     const int32_t    doc_id      = posting_ivars->doc_id;
-    const uint32_t   delta_doc   = doc_id - ivars->last_doc_id;
+    const uint32_t   delta_doc   = (uint32_t)(doc_id - ivars->last_doc_id);
     char  *const     aux_content = posting_ivars->blob
                                    + posting_ivars->content_len;
     if (posting_ivars->freq == 1) {
